Function size() returns size_t when std::vector is used.
I am thinking what would happe in std::vector would be replaced with QVector, 
and std::string with QString? Their size() function return int, and then
there would be less to think about unsigned int vs. signed conversions.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10791 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/base/Studio.cpp
@@ -157,7 +157,7 @@ Studio::getSpareDeviceId(InstrumentId &baseInstrumentId)
 	ids.insert((*it)->getId());
 	if ((*it)->getType() == Device::Midi) {
 	    InstrumentList il = (*it)->getAllInstruments();
-	    for (int i = 0; i < il.size(); ++i) {
+	    for (size_t i = 0; i < il.size(); ++i) {
 		if (il[i]->getId() > highestMidiInstrumentId) {
 		    highestMidiInstrumentId = il[i]->getId();
 		    foundInstrument = true;
